refactor(debugging): add LogSignal subclass (#5760)
The LogSignal subclass is used for signals that have the need to emit
log messages. Typical examples are for when a log message is to be
emitted (as part of the log probe triggering, or when condition errors
have to be reported). This decouples log messages from just snapshot
signals and makes them available to potentially any probe type.