ci: unblock ci by setting typing extensions to latest (#5949)
## Background

We specify typing_extensions in .riot/requirements/....txt files but
this package is overridden by the base virtual env generated by riot.
Normally this is not an issue but this leafs to module not found errors
when running tests subprocesses. The underlying issue is still unknown.
This could be an issue with riot, how we execute subprocess tests, how
we set the pythonpath, or none of the above.

## Impact

Although this unblocks ci for now, this issue will resurface if there is
another release of typing_extensions. If there is another release of
typing extensions will need to bump the version once again to unblock ci
but this process is less than ideal.


## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).

## Reviewer Checklist

- [ ] Title is accurate.
- [ ] No unnecessary changes are introduced.
- [ ] Description motivates each change.
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [ ] Testing strategy adequately addresses listed risk(s).
- [ ] Change is maintainable (easy to change, telemetry, documentation).
- [ ] Release note makes sense to a user of the library.
- [ ] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.