fix(profiling): handle partially loaded threading module (#5847)
When using `gevent` the application needs to reload some modules that
are also used by `ddtrace`. One such modules is `threading`, which plays
an important role in the current implementation of the profiler. Since
the stack collector needs to access both the `ddtrace` and the
application copy of the `threading` module, with the profiler started
before the new copy is created by the application, there is a race
between the stack sampling and the module import that can result in the
stack collector trying to access attributes from the new copy of
`threading` before its import is completed.

As a fix for this issue, we guard against expected `AttributeError`s.

## Testing strategy

An existing gunicorn + gevent test has been extended to catch any
potential signs of the issue occurring again in the future. Given that
this is a race condition, the issue itself is hard to reproduce
systematically.

## Risks

We don't expect any risks in guarding for well-known and stable
attributes of the `threading` module.

---------

Co-authored-by: sanchda <sanchda@users.noreply.github.com>
Co-authored-by: Gabriele N. Tornetta <gabriele.tornetta@datadoghq.com>
Co-authored-by: Gabriele N. Tornetta <P403n1x87@users.noreply.github.com>