chore(peer.service) - peer.service logic + flag
== Special Notes ==
These changes are gated behind a feature flag and are not expected to be
stable until a release note is generated. Existing customers should be
unaffected.

The logic relies on 'precursor' tags, which are being added in other PRs.
Logic will not be fully enabled until until the precursors are
completed.

These changes are reliant on service naming as well, and will not be
fully enabled until that initiative is complete.

== Peer.Service Summary ==
The changes for peer.service will enable new functionality in the
service map, when enabled, in conjunction with service naming.

The changes in this PR are:

1. Introduce the peer.service computation, gated behind a feature flag
2. Add in the peer.service and peer.service.source computation logic

When peer service is enabled, the logic is applied as a post-processor
on spans, as part of the "on_finish" methods
enabled.