fix: avoid importing asyncio
Whilst the library has many integrations and features relying on
asyncio, the actual products don't make use of it. The interaction
is only aimed at providing observability into Python's native async
support. We then avoid importing from asyncio and rely only on
import hooks to achieve our goals. This prevents the asyncio module
from being loaded, which also means that we no longer need to unload
it as part of the better support for frameworks such as gevent.