chore(telemetry): fix interval set on telemetry metrics (#6025)
Includes the following changes:
- An interval should only be set for gauge and rate metrics. Any
interval set on count and distribution metrics will be ignored by the
backend. This is not a "fix", it more of a "cleanup".
- The interval set on gauge and rate metrics is incorrect. This PR
ensures the correct value is used (PeriodicService.interval should be
used instead of ``DD_TELEMETRY_HEARTBEAT_INTERVAL``).

This is a fix for ddwaf metrics that are used internally. Since this fix
does not impact users, a release note is not required.


## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.