refactor: drop tenacity dependency (#5998)
We implement a simple retry logic internally to drop the dependency on
tenacity. This will reduce the number of modules that we import
indirectly through dependencies. In particular, with this change we
avoid importing asyncio. This is to avoid issues like #5989, which are
caused by the module cloning solution introduced for better supporting
frameworks like gevent.