Make procfs-based metrics lookup more tolerant
It turns out that some managed environments, such as Google Cloud Run,
are implemented in such a way that procfs status lines are withheld even
though the underlying kernel version would indicate that those lines
should be available.

This patch attempts to make reading procfs resiliant to this behavior,
in an attempt to make dd-trace-py available to Google Cloud Run users.