chore(peer.service): peer.service logic + flag (#6042)
== Special Notes ==
These changes are gated behind a feature flag and are not expected to be
stable until a release note is generated. Existing customers should be
unaffected.

The logic relies on 'precursor' tags, which are being added in other
PRs. Logic will not be fully enabled until until the precursors are
completed.

These changes are reliant on service naming as well, and will not be
fully enabled until that initiative is complete.

== Peer.Service Summary ==
The changes for peer.service will enable new functionality in the
service map, when enabled, in conjunction with service naming.

The changes in this PR are:

1. Introduce the peer.service computation, gated behind a feature flag
2. Add in the peer.service and peer.service.source computation logic

When peer service is enabled, the logic is applied as a post-processor
on spans, as part of the "on_finish" methods
enabled.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).

## Reviewer Checklist

- [ ] Title is accurate.
- [ ] No unnecessary changes are introduced.
- [ ] Description motivates each change.
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [ ] Testing strategy adequately addresses listed risk(s).
- [ ] Change is maintainable (easy to change, telemetry, documentation).
- [ ] Release note makes sense to a user of the library.
- [ ] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.