chore(tracing): adds `out.host` tag for consul (#6076)
Adds `out.host` tag for consul as a precursor tag for `peer.service`. 
Testing is updated to check for new tag 

The larger initiative at play here is to set the `peer.service` tag
which depends on a group of precursor tags to extract information from.
`out.host` is one of those precursor tags which must be set in certain
target integrations of which consul is one. Using this information, we
can accurately set the `peer.service` tag which will allow for future
development with accurate hostname information.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.

Co-authored-by: Munir Abdinur <munir.abdinur@datadoghq.com>