test: fix parametrisation support in subprocess marker (#6085)
A bug in the handling of the parametrize argument to the subprocess
marker prevented it from generating the correct test matrix. This fix
ensures that the correct matrix dictionary is generated so that all the
intended test cases are actually run.

(cherry picked from commit 5d174c8892550798d497597f6c6cbbab922d7410)