chore(peer.service): add peer.service remap logic
Peer.Service is currently a beta product and is hidden behind a feature
flag, alongside service naming.

This change provides users with the ability to change peer.service names
via an environment variable, `DD_TRACE_PEER_SERVICE_MAPPING`. When users
specify this variable AND peer.service is enabled, any services in the
mapping will be automatically remapped to a different value, and the tag
`_dd.peer.service.remapped_from` will be set to the prior value.