fix(pymemcache): allow hashclient integration to override pin when use_pooling=True (#6084)
This pull request fixes
https://github.com/DataDog/dd-trace-py/issues/5637 by putting a new
`ObjectProxy` wrapper around pymemcache's `HashClient`.

This wrapper proxies its command invocations to the underlying
`HashClient` instance. When the `use_pooling` setting is in use, this
wrapper starts a span before doing the proxy call. This is necessary
because the `use_pooling` setting causes `Client` instances to be
created and destroyed dynamically in a manner that isn't affected by the
`patch()` function.

The implementation shares as much code as possible while maintaining
readability. It also adds a test for the specific use case described in
#5637.