chore(peer.service): report hostname for elasticsearch
The peer.service initiative allows Datadog to set better defaults
for service names, and to build a more intuitive service map, by mapping
outgoing connections with the destinations service names.  In order to
implement this functionality, various 'precursor' tags must be added to

This change adds one such precursor tag required for peer.service, for
elasticache, which allows us to identify the cluster being connected to.

This change adds the 'out.host' tag for elasticsearch spans using the
first connection which returns a valid hostname. If none are found, the
tag is not added.