fix(runtime): don't throw exceptions when procfs is filtered (#6077)
Some managed environment, such as Google Cloud Run, filter the view of
`/proc/<pid>/status` so that only a subset of the status lines (which
would ordinarily be available given the underlying kernel version) are
presented.

This patch zeroes any metric which cannot be read for whatever reason.

I also went ahead and slightly generalized the mechanism we used to
register these runtime metrics. I was responding to a TODO comment in
the implementation I found agreeable, but this aspect of the PR isn't
necessary for the fix.

In pathological cases where procfs opens, but we're missing required
status lines, the runtime will be spending additional time handling the
exceptions (which are not free), but I anticipate that:
* This scenario is comparatively minor
* The rate at which we harvest metrics is low enough that this won't
matter.

I'm not offering any novel tests.  :shrug:

## Checklist

- [X] Change(s) are motivated and described in the PR description.
- [X] Testing strategy is described if automated tests are not included
in the PR.
- [X] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [X] Change is maintainable (easy to change, telemetry, documentation).
- [X] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [X] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).

## Reviewer Checklist

- [ ] Title is accurate.
- [ ] No unnecessary changes are introduced.
- [ ] Description motivates each change.
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [ ] Testing strategy adequately addresses listed risk(s).
- [ ] Change is maintainable (easy to change, telemetry, documentation).
- [ ] Release note makes sense to a user of the library.
- [ ] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.