chore(tracing): surface rpc.service in grpc contrib (#5946)
## Motivation

For upcoming computation of peer service, we want to have the
`rpc.service` precursor tag available in GRPC client spans. This tag is
the combination of the GRPC package name and the GRPC service name (for
example, when emitting a span for a call to the full path
`/myPackage.myService/myMethod`, the value of `rpc.service` is
`myPackage.myService`).
The tag name is defined in the separate effort for consistent span tags
across tracers.

## Change

Surface `rpc.service` in GRPC contrib. The package name may be missing
in some cases, and in these cases the value of this tag is _only_ the
GRPC service name.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).

## Reviewer Checklist

- [ ] Title is accurate.
- [ ] No unnecessary changes are introduced.
- [ ] Description motivates each change.
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [ ] Testing strategy adequately addresses listed risk(s).
- [ ] Change is maintainable (easy to change, telemetry, documentation).
- [ ] Release note makes sense to a user of the library.
- [ ] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.

---------

Co-authored-by: Zachary Groves <32471391+ZStriker19@users.noreply.github.com>