chore(flaky test): fix flaky integration tests
This change updates the 'integration' test suite to ensure spans fall
into the same bucket and aren't transmitted to the test agent via
separate invocations of the period processor.

This fixes a set of flaky tests where the number of buckets was
unexpected due to grouping spans due to either separate time buckets or
across invocations of the stats span processor.