refactor(config): add support for config sources, subscription
With the current global config there are several issues:

There is no way to distinguish between user set overrides, environment settings or
defaults.

No way to subscribe to global config changes. This leads to the global
config having to be read each time in case the user defines an override
or the global config is read once at start-up and cannot be modified
again.

Consistency between configuration options. Documentation, testing, etc.
Whether there is an in-code API or not.

Mutability of in-code configuration items is undefined.

To address 1) and 3) the global config is extended with declarative
config items. Config items provide a way to declare environment
variables, defaults, documentation and other metadata. By default,
config items resolve to the value to be the user override, if provided,
else the environment variable, if provided, else the default. The
resolution logic can be overridden per config item. Config items can
return their configuration source which can be used for debugging.

For 2) and 4) the global config is extended to support subscribing to
items. This enables objects to subscribe to changes to the global config
rather than having to poll it or lock it in at start-up.

This change is backwards compatible with the current global config and does not introduce any new public API surface area.

The main risk of this change is a performance regression in looking up
config items (just config.service now). There is no optimization done to
the lookup so for each time the service setting is requested, it is
possible that the the user setting and env vars are checked each time.