feat(debugging): exception debugging (#5981)
Add the feature of capturing debug information along the traceback of
any exceptions attached to spans within traces. Whenever a span is
marked with an error, we retrieve the exception information, unwind the
chain, and take snapshots of all the frames referenced by the traceback.
This way we can enhance the error tracking UI with information about
local variables/function arguments to provide more context about
exceptions along traces.