chore(debugging): skip stdlib, third-party frames from exceptions
When collection data from an exception attached to a trace in
exception debugging, we make the decision to skip frames that are
associated to the standard library or third-party libraries, and
only capture debugging information for user code. We do so by
leveraging the code provenance data, together with the extra
information that comes from the DD_MAIN_PACKAGE environment variable
of the git metadata embedding support.

Note that when the variable is not set, we treat every non-stdlib
path as user code to avoid not capturing any data at all.