fix(profiling): handle known HTTP request errors (#6249)
After the removal of the tenacity dependency, some known HTTP request
exceptions are no longer handled as expected and the user is asked to
report an issue to the project. This change ensures that we have a
similar behaviour in place whereby known issues are handled as before.

(cherry picked from commit e0972224c6a4f249fe6a85ab22428e8174c66818)