chore: move constants to internal module (#6299)
This change moves many of the constants used in IAST taint tracking to
`ddtrace.internal.constants`. We're moving toward a more clear
separation of concerns between Products like AppSec and Contrib
integrations like Flask and Django.

In the case of request blocking, the separation of concerns ideally
breaks down as follows. The AppSec Product code in the `ddtrace/appsec`
directory knows how to make a block/don't block decision based on
communication with libddwaf. The Django code in `ddtrace/contrib` knows
how to take that blocking decision into account when processing
requests.

According to this separation of concerns, these constants will be used
by both the Product and the Contrib. Moving them to `internal.constants`
means that the Contrib won't have to talk about AppSec when it imports
them.