chore(flask): remove appsec code from flask contrib (#6333)
This change much of the non-IAST Application Security code from
`contrib.flask.patch` to `appsec.handler`, clarifying the separation of
concerns between the two. It accomplishes this by using the API at
`ddtrace.internal.core`.

In the case of request blocking, the separation of concerns ideally
breaks down as follows. The AppSec Product code in the `ddtrace/appsec`
directory knows how to make a block/don't block decision based on
communication with libddwaf. The Flask code in `ddtrace/contrib` knows
how to take that blocking decision into account when processing
requests.

Co-authored-by: Federico Mon <federico.mon@datadoghq.com>
Co-authored-by: Alberto Vara <alberto.vara@datadoghq.com>
Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>