chore(profiling): add pure-Python traceback backend
We introduce a private environment variable to control the backend
used for unwinding Python stacks. By default the backend is set to
Cython, which is what was previously used. The variable can be set
to `py` to allow switching to a pure-Python traceback backend in
case of potential issues with the Cython backend.