ci: run the debugger suite only if necessary (#6412)
We introduce the concept of suitespec as a way of describing how sources
affect test runs. We use it to ensure that the debugger tests run only
if sources that the suite depends on are modified by the current commit.

## Suitespec Implementation Details

The suitespec solution is based on a manual configuration of of test
suites. To simplify the declaration of file patterns for test suites,
one can make use of _components_, which essentially are a logic
collection of patterns. Test suite can then be declared as a list of
components to reflect their dependencies on these logic parts, and to
DRY the declaration itself by avoiding repetitions.

## Notes

- When the script fails for any reason, tests are run.
- It is important that path patterns are listed correctly, or some tests
might not run when they are in fact supposed to.
- Best effort to determine the correct list of changed files via the
GitHub REST API. When that fails, we fall back to the less accurate `git
diff` against the target branch.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [ ] Title is accurate.
- [ ] No unnecessary changes are introduced.
- [ ] Description motivates each change.
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [ ] Testing strategy adequately addresses listed risk(s).
- [ ] Change is maintainable (easy to change, telemetry, documentation).
- [ ] Release note makes sense to a user of the library.
- [ ] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [ ] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)