chore(flask): remove IAST code from Flask contrib (#6334)
This change removes the IAST-related code from `contrib.flask.patch`,
clarifying the separation of concerns. It accomplishes this by using the
API at `ddtrace.internal.core`.

In the case of request blocking, the separation of concerns ideally
breaks down as follows. The AppSec Product code in the `ddtrace/appsec`
directory knows how to make a block/don't block decision based on
communication with libddwaf. The Wsgi code in `ddtrace/contrib` knows
how to take that blocking decision into account when processing
requests.

---------

Co-authored-by: Federico Mon <federico.mon@datadoghq.com>
Co-authored-by: Alberto Vara <alberto.vara@datadoghq.com>
Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>